refactor: LotteryCalculator totalProfit()를 WinningResult에게 위임
