+++ src/test/java/lotterymachine/utils/LotteryCalculatorTest.java
@@ -29,22 +29,5 @@ public class LotteryCalculatorTest {
         assertThat(result).isEqualTo(0.06);
     }
 
-    @Test
-    @DisplayName("로또 결과를 입력 받아, 수익을 계산하여 반환한다.")
-    void totalProfit() {
-        List<LotteryNumber> input = IntStream.range(1, 7)
-                .mapToObj(LotteryNumber::valueOf)
-                .collect(Collectors.toList());
-        List<LotteryTicket> tickets = List.of(new LotteryTicket(input));
-        LotteryTickets lotteryTickets = new LotteryTickets(tickets);
 
-
-        List<LotteryNumber> input2 = IntStream.range(4, 10)
-                .mapToObj(LotteryNumber::valueOf)
-                .collect(Collectors.toList());
-        WinningLottery winningLottery = new WinningLottery(input2, LotteryNumber.valueOf(12));
-        WinningResult winningResult = new WinningResult(lotteryTickets, winningLottery);
-        int result = LotteryCalculator.totalProfit(winningResult.getResult());
-        assertThat(result).isEqualTo(5000);
-    }
 }
